+++ src/main/java/nextstep/subway/line/dto/PathRequest.java
@@ -1,0 +1,19 @@
+package nextstep.subway.line.dto;
+
+public class PathRequest {
+    private final Long source;
+    private final Long target;
+
+    public PathRequest(Long source, Long target) {
+        this.source = source;
+        this.target = target;
+    }
+
+    public Long getSource() {
+        return source;
+    }
+
+    public Long getTarget() {
+        return target;
+    }
+}
