+++ src/main/java/nextstep/subway/line/dto/PathResponse.java
@@ -1,0 +1,16 @@
+package nextstep.subway.line.dto;
+
+import java.util.List;
+import nextstep.subway.station.dto.StationResponse;
+
+public class PathResponse {
+
+    private List<StationResponse> stations;
+    private int distance;
+
+    public PathResponse(List<StationResponse> stations, int distance) {
+        this.stations = stations;
+        this.distance = distance;
+    }
+
+}
