+++ src/main/java/nextstep/subway/line/ui/PathController.java
@@ -1,9 +1,9 @@
-package nextstep.subway.path.ui;
+package nextstep.subway.line.ui;
 
 
-import nextstep.subway.path.application.PathService;
-import nextstep.subway.path.dto.PathRequest;
-import nextstep.subway.path.dto.PathResponse;
+import nextstep.subway.line.application.PathService;
+import nextstep.subway.line.dto.PathRequest;
+import nextstep.subway.line.dto.PathResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
