--- src/main/java/nextstep/subway/path/dto/PathRequest.java
@@ -1,11 +1,0 @@
-package nextstep.subway.path.dto;
-
-public class PathRequest {
-    private Long source;
-    private Long target;
-
-    public PathRequest(Long source, Long target) {
-        this.source = source;
-        this.target = target;
-    }
-}
