--- src/main/java/nextstep/subway/path/dto/PathResponse.java
@@ -1,10 +1,0 @@
-package nextstep.subway.path.dto;
-
-import java.util.List;
-import nextstep.subway.station.dto.StationResponse;
-
-public class PathResponse {
-
-    private List<StationResponse> stations;
-    private int distance;
-}
